add active select on menu
