+++ client/src/screens/Admin/Layout/index.module.css.d.ts
@@ -1,0 +1,7 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'selected': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
